+++ src/components/App/components/public/Connectors/Connector/ListView/index.js
@@ -7,6 +7,7 @@ import EditorComponent from '../../../..';
 
 import Iterable from "./Iterable";
 import NamedField from '../Fields/NamedField';
+import DefaultValue from '../Fields/NamedField/DefaultValue';
 
 export const ObjectContext = createContext({});
 
@@ -137,7 +138,7 @@ class ListView extends EditorComponent {
           return null;
         }
 
-        const children = super.renderChildren();
+        let children = super.renderChildren() || [];
 
         let output = null;
 
@@ -172,22 +173,32 @@ class ListView extends EditorComponent {
         }
         else {
 
-          output = items.length ? items.map((n, index) => {
+          if (items.length) {
 
-            const {
-              id,
-            } = n;
+            children = children.filter(n => n && n.type !== DefaultValue);
 
-            return <ObjectContext.Provider
-              key={id || index}
-              value={{
-                object: n,
-                ...otherData,
-              }}
-            >
-              {children}
-            </ObjectContext.Provider>
-          }) : null;
+            output = items.map((n, index) => {
+
+              const {
+                id,
+              } = n;
+
+              return <ObjectContext.Provider
+                key={id || index}
+                value={{
+                  object: n,
+                  ...otherData,
+                }}
+              >
+                {children}
+              </ObjectContext.Provider>
+            });
+
+          }
+          else {
+            output = children.filter(n => n && n.type === DefaultValue);
+          }
+
 
           {/* output = <Grid
             container